fix : store 구조 변경
